trouble with event helpers, they need access to state, also adjusted visual route path of events page to just say /meetups
